+++ src/pupil_labs/realtime_api/__init__.py
@@ -27,7 +27,6 @@ except PackageNotFoundError:
 
 __all__ = [
     "__version__",
-    "__version_info__",
     "APIPath",
     "Device",
     "DeviceError",